Work around defect in gcc 4.8.5
GCC 4.8.5 fails to handle explicit template specializations correctly,
resulting in an infinite loop of the specialization calling itself,
rather than resolving to the template.
